Reuse of probe identifiers and change of outgoing source ports
The client now reuses answered probe identifiers of reverse traceroute
messages to reuse existing NAPT mapping (as the probe identifier
is encoded into the ICMP identifier, each new probe identifiers
results a new NAPT mapping). This change will reduce the amount
of state kept in NAPT middleboxes for reverse-traceroute execution.

The outgoing source port for classic traceroute probes has been changed
from the experimental port 1021 to the IANA allocted port for
traceroute: 33434
This change is introduced to avoid possible conflicts when running
the client on the same machine as the reverse-traceroute server,
which still uses port 1021 and may drop incoming traffic to that port.