Updated probe response handling
When possible probe responses were received but no
associated session entry was present, we simply passed
the packet onward.

As for both UDP and TCP a reserved port should be assigned for
reverse traceroute probes, such probe responses are now dropped
even with no session present.

ICMP is a special case: As we cannot reserve the ID field in the
ICMP header, as it is frequently used by classic traceroute/ping
to match request and responses, the identifier of such programs
may collide with reverse traceroute probes.

To mitigate the effects of such rare collisions, any possible ICMP probe responses
are passed on if no session exists, which allows classic traceroute/ping
to parse those packets.