Updated ICMP probe encoding
Until now, the ICMP identifier of the probe contained the constant probe
id, which is used to check if the response may be an answer to a probe.
This has the drawback, that when ping/traceroute are run on the server
machine, the ICMP identifier could overlap, leading to ping/traceroutes
responses being dropped by reverse traceroute.

Now the constant probe identifier is encoded into the ICMP sequence,
setting it to the constant value 0xffff.
The clients query id is encoded into the probes identifier.
This the following advantage:

As classic traceroute uses its process id to encode it inside the
ICMP identifier of generated probes and sequentially increments the
ICMP sequence field to match individual probes with responses,
the likelyhood of clashes with a reverse traceroute server is
minimal: That is because we chose the highest possible value for the
probe identifier to be encoded into the ICMP sequence field. Classic
traceroute/ping will terminate before they reach such a high
sequence, and even if they do, this will lead to a single response
being dropped by the interference of reverse traceroute.

Paris-Traceroute on the other hand increments the ICMP identifier
and keeps the sequence at 0. There should be no clashes with
this tool either.