Introduced restriction for source networks
The server can now be configured to allow reverse-traceroute requests
from a set of preconfigured networks. The loader reads them from a file
with the --allow-sources-from argument.
Each line in the file should contain a network in CIDR notation.
Each entry is validated and invalid entries are skipped by the loader.