Added debian packaging support for client
The client can now be built into a binary executable.
The main logic is inside scripts/binarize.sh and pulls a
new copy of the poetry environment with pyinstaller as a dev-dependency.
PyInstaller is then responsible to create the binary in question.
A new Makefile allows the debhelper commands to correctly package the result.