Introduced global session ID to server
As the previous session key consisted of the traceroute target
and the client provided identifier, multiple clients could
create the same session key on the server, resulting in ambiguity.

Now each client-supplied identifier is converted into an identifier
unique to the target. The client-local identifier is persisted in the state object.
This means that each target can contain at most 2^16 sessions.
If the buffer is full, the request is dropped.