Backported session mapping to kernel version 5.15
As the target-specific mapping (using the bpf_loop helper) requires
at least kernel 5.17, the implementation was backported to another (less elegant)
solution: Using a global queue of usable stack identifiers (not target-specific).
The queue is populated by the program loader, the eBPF program then pushes and
pops new and used identifiers as needed. This approach is more memory intensive
and limits the maximum sessions (globally) to 65535.

When kernel 5.17 support is more widely adopted the target-specific mapping
via bpf_loop should be reintroduced, as it does not suffer from said limitation.