Removed possible race condition between map_lookup and map_delete
When receiving multiple identical probe responses (should not happen)
with matching session key, a race condition between session lookup
and removal was possible, leading to a duplicate global identifier returned
to the queue.
Now lookup and delete are merged into a seperate function that
returns an error if the atomic delete operation failed and only
returns an identifier to the queue on success.