Fixed bugs related to ethernet padding
When the traceroute request is smaller than the minimum ethernet payload size it is padded.
The skb->len field includes the size of this padding.
When checking if we reached the end of the packet we do not want to consider the padding, hence we have to rely on the payload size as reported by the IP header.